requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
gspread
oauth2client
python-dotenv
requests
python-multipart
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary

